backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.9
mysqlclient==2.1.1
mysql-connector-python==8.0.33
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==1.10.7
email-validator==2.0.0
python-dotenv==1.0.0
